+++ projects/rocprofiler-compute/src/rocprof_compute_analyze/analysis_db.py
@@ -20,7 +20,6 @@ from utils.analysis_orm import Database
 from utils.file_io import load_pc_sampling_results, process_pc_sampling_kernel_trace
 from utils.logger import (
     console_debug,
-    console_error,
     console_warning,
     demarcate,
 )
@@ -57,7 +56,7 @@ from utils.utils_analysis import (
     PEAK_COL_PREFERENCE,
     VALUE_COL_PREFERENCE,
 )
-from utils.utils_common import _PROFILE_OUTPUT_FORMAT, get_uuid, get_version
+from utils.utils_common import get_uuid, get_version
 from utils.utils_counter_defs import extract_counters_and_variables, get_build_in_vars
 
 
@@ -85,14 +84,6 @@ class db_analysis(OmniAnalyze_Base):
     def pre_processing(self) -> None:
         """Perform any pre-processing steps prior to analysis."""
         super().pre_processing()
-        if (
-            self._profiling_config.get("format_rocprof_output")
-            != _PROFILE_OUTPUT_FORMAT
-        ):
-            console_error(
-                "Creation of analysis database is only supported "
-                "for profiling data with rocpd output format."
-            )
 
         self._roofline_ceilings_per_workload = self.calc_roofline_ceilings()
         pc_sampling_tool_data = (